Made funtionality for player and team

+++ src/main/java/polishchuk/repository/PlayerRepository.java
@@ -7,7 +7,7 @@ import polishchuk.entity.Player;
 import java.util.Optional;
 
 @Repository
-public interface PlayerRepository extends JpaRepository<Player, Long> {
+public interface PlayerRepository extends JpaRepository<Player, Integer> {
 
     public Optional<Player> findByLastName(String lastName);
